SDLNAProject/requirements.txt
alabaster==0.7.8
APScheduler==3.6.3
apsw==3.28.0.post1
apturl==0.5.2
arandr==0.1.10
astroid==2.3.3
async-generator==1.10
attrs==19.3.0
Babel==2.6.0
backcall==0.1.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.4
Brlapi==0.7.0
ccsm==0.9.14.1
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.9
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
Click==7.0
cloudpickle==1.3.0
cmdtest==0.32+git
colorama==0.4.3
command-not-found==0.3
compizconfig-python==0.9.14.1
cryptography==2.8
css-parser==1.0.4
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
delta-spark==1.1.0
distro-info==0.23ubuntu1
distro==1.4.0
dnspython==1.16.0
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fasteners==0.14.1
feedparser==5.2.1
future==0.18.2
fuzzywuzzy==0.18.0
h11==0.13.0
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
imagesize==1.2.0
importlib-metadata==1.5.0
ipykernel==5.2.0
ipython-genutils==0.2.0
ipython==7.13.0
isort==4.3.4
jarowinkler==1.0.2
jedi==0.15.2
Jinja2==2.10.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
Levenshtein==0.18.1
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
lxml==4.7.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mccabe==0.6.1
mechanize==0.4.5
mistune==0.8.4
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
msgpack==0.6.2
multitasking==0.0.10
mysql-connector-python==8.0.27
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
nose==1.3.7
numpy==1.21.5
numpydoc==0.7.0
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.9
outcome==1.1.0
packaging==20.3
pandas==1.3.5
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.5.2
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pip==20.0.2
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.5.1
py4j==0.10.9.3
pyarrow==7.0.0
PyBluez==0.22
pycairo==1.16.2
pychm==0.8.6
pycodestyle==2.5.0
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==2.4.6
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
pyspark==3.2.1
pysvn==1.9.9
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian==0.1.36ubuntu1
pytz-deprecation-shim==0.1.0.post0
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
QtAwesome==0.4.4
qtconsole==4.6.0
QtPy==1.9.0
rapidfuzz==2.0.7
regex==2019.8.19
reportlab==3.5.34
repoze.lru==0.7
requests-unixsocket==0.2.0
requests==2.22.0
roman==2.0.0
rope==0.10.5
Routes==2.4.1
SecretStorage==2.3.1
selenium==4.1.3
setuptools==45.2.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==1.9.5
Sphinx==1.8.5
spyder-kernels==0.5.2
spyder==3.3.6
ssh-import-id==5.10
system-service==0.3
systemd-python==234
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.3
trio-websocket==0.9.2
trio==0.20.0
ttystatus==0.38
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
WebOb==1.8.5
wheel==0.34.2
wrapt==1.11.2
wsproto==1.1.0
xkit==0.0.0
xlrd==2.0.1
yfinance==0.1.68
youtube-dl==2020.3.24
zeroconf==0.24.4
zipp==1.0.0
